requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
chainlit>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.0.0
